feat: complete process sheet enhancement with internal confirm, PDF print, contract logs
- Add internal confirm user configuration (process-sheet-level confirm_user_ids,
  system-level default via WeCom settings, customer confirm auto-notify)
- Add contract change-diff logging on create/edit with detailed field comparison
- Add cancel/restore workflow with contract version snapshot
- Rewrite PDF print with professional 工艺确认单 format, fix data source
  (use ProcessSheetItem instead of ContractItem), add permission check
- Add StatusLog shared component for unified log display
- Add version timeline with customer confirm events to SheetDetail
- Fix reopen-edit to keep version number unchanged
- Remove SheetList create button (only push-down from contract items)
- Add process_sheet_id index to production_log model
- Add lazily-loaded WeasyPrint to avoid startup crashes on missing system libs

+++ backend/app/schemas/contract_item.py
@@ -104,7 +104,7 @@ class ContractItemOut(BaseModel):
     process_sheet_id: Optional[int] = None
     process_sheet_no: Optional[str] = None
     process_sheet_status: Optional[str] = None
-    process_sheet_version: Optional[int] = 0
+    process_sheet_version: Optional[float] = 0
 
     class Config:
         from_attributes = True
